Created: chat-item.tsx, deleted-message.tsx, [messageId], & renamed messages.ts to index.ts
Updated:
page.tsx,
chat-messages.tsx,
modal-provider,
use-modal-store,
io.ts.

We debugged an issue with socketUrl being typo as sockerUrl which was
causing a 404,

an issue with imports not called correctly where
ServerIO was changed to ServerIo so that the import from socket.io
could respond.

Last thing:
we changed the next.config file to reflect more simple routing
structure until
we need to start customizing routes for safety &
updated the POST route as well as debugged a PATCH route.